+++ app/src/main/kotlin/org/fossify/notes/adapters/NotesPagerAdapter.kt
@@ -96,6 +96,10 @@ class NotesPagerAdapter(fm: FragmentManager, val notes: List<Note>, val activity
         (fragments[position] as? TasksFragment)?.removeCheckedItems()
     }
 
+    fun uncheckAllItems(position: Int) {
+        (fragments[position] as? TasksFragment)?.uncheckAllItems()
+    }
+
     fun refreshChecklist(position: Int) {
         (fragments[position] as? TasksFragment)?.saveAndReload()
     }