+++ CHANGELOG.md
@@ -5,9 +5,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [1.4.0] - 2025-07-15
 ### Added
 - Option to delete the last open note ([#157])
 - Option to uncheck all checked items ([#156])
+
+### Changed
+- Updated translations
 
 ## [1.3.1] - 2025-07-12
 ### Changed
@@ -69,7 +74,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#164]: https://github.com/FossifyOrg/Notes/issues/164
 [#178]: https://github.com/FossifyOrg/Notes/issues/178
 
-[Unreleased]: https://github.com/FossifyOrg/Notes/compare/1.3.1...HEAD
+[Unreleased]: https://github.com/FossifyOrg/Notes/compare/1.4.0...HEAD
+[1.4.0]: https://github.com/FossifyOrg/Notes/compare/1.3.1...1.4.0
 [1.3.1]: https://github.com/FossifyOrg/Notes/compare/1.3.0...1.3.1
 [1.3.0]: https://github.com/FossifyOrg/Notes/compare/1.2.0...1.3.0
 [1.2.0]: https://github.com/FossifyOrg/Notes/compare/1.1.1...1.2.0